:truck: move: 필요한 페이지 생성 및 폴더 구조화
