+++ src/pages/MyPage/index.js
@@ -2,8 +2,8 @@ import React from "react";
 import styled from "styled-components";
 import NavBar from "../../components/NavBar";
 import AppBar from "../../components/AppBar";
-import ProfilePic from "./ProfilePic";
-import SelectBtn1 from "./SelectBtn1";
+import ProfilePic from "./components/ProfilePic";
+import SelectBtn1 from "./components/SelectBtn1";
 
 const MyPage = () => {
     return (
